Day24.java

+++ Day24.java
@@ -2,10 +2,11 @@ public class Day24 {
     public static void main(String[] args) {
         // Latihan:Program luas lingkaran
         final double PHI = 3.14; 
-        double r = 5;
+        int r = 5;
         double luas = PHI * r * r;
 
         System.out.println("Luas lingkaran: " + luas) ;
     }
 
 }
+
